Refactored location of everything so that client-side JS and server-side JS are separate
