+++ JS-OCR-demo-gh-pages/public/css/Jcrop.gif
