+++ JS-OCR-demo-gh-pages/public/css/main.css
