+++ JS-OCR-demo-gh-pages/public/fonts/glyphicons-halflings-regular.ttf
