+++ JS-OCR-demo-gh-pages/public/img/process.xcf
