+++ JS-OCR-demo-gh-pages/public/img/step1.png
