+++ JS-OCR-demo-gh-pages/public/img/step2.png
