+++ JS-OCR-demo-gh-pages/public/js/main.js
