+++ JS-OCR-demo-gh-pages/realtimetextreader.js
