+++ JS-OCR-demo-gh-pages/views/layouts/index.hbs
@@ -1,0 +1,33 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+
+    <title>JavaScript OCR demo</title>
+
+    <meta name="Description" content="Optical Character Recognition demo in JavaScript"/>
+    <meta property="og:image" content="http://kdzwinel.github.io/JS-OCR-demo/img/process.png"/>
+    <link rel="image_src" href="http://kdzwinel.github.io/JS-OCR-demo/img/process.png"/>
+
+    <link rel="stylesheet" href="/css/bootstrap.min.css"/>
+    <link rel="stylesheet" href="/css/jquery.Jcrop.min.css"/>
+    <link rel="stylesheet" href="/css/main.css"/>
+    <link href="https://netdna.bootstrapcdn.com/font-awesome/4.0.3/css/font-awesome.css" rel="stylesheet">
+  </head>
+  <body>
+    {{{body}}}
+
+
+
+    <script src="https://code.jquery.com/jquery-1.10.2.min.js"></script>
+    <script src="/js/vendor/modernizr.min.js"></script>
+    <script src="/js/vendor/bootstrap.min.js"></script>
+    <script src="/js/vendor/jquery.Jcrop.js"></script>
+    <script src='https://cdn.jsdelivr.net/gh/naptha/tesseract.js@v1.0.14/dist/tesseract.min.js'></script>
+    <script src="/js/vendor/glfx.min.js"></script>
+
+    <script src="/js/main.js"></script>
+  </body>
+</html>
