+++ JS-OCR-demo-gh-pages/views/main.hbs
@@ -1,36 +1,3 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-
-    <title>JavaScript OCR demo</title>
-
-    <meta name="Description" content="Optical Character Recognition demo in JavaScript"/>
-    <meta property="og:image" content="http://kdzwinel.github.io/JS-OCR-demo/img/process.png"/>
-    <link rel="image_src" href="http://kdzwinel.github.io/JS-OCR-demo/img/process.png"/>
-
-    <link rel="stylesheet" href="css/bootstrap.min.css"/>
-    <link rel="stylesheet" href="css/jquery.Jcrop.min.css"/>
-    <link rel="stylesheet" href="css/main.css"/>
-    <link href="https://netdna.bootstrapcdn.com/font-awesome/4.0.3/css/font-awesome.css" rel="stylesheet">
-    
-    <script type="text/javascript">
-        var _gaq = _gaq || [];
-        _gaq.push(['_setAccount', 'UA-145083566-1']);
-        _gaq.push(['_trackPageview']);
-
-        (function() {
-        var ga = document.createElement('script'); ga.type = 'text/javascript'; ga.async = true;
-        ga.src = ('https:' == document.location.protocol ? 'https://ssl' : 'http://www') + '.google-analytics.com/ga.js';
-        var s = document.getElementsByTagName('script')[0]; s.parentNode.insertBefore(ga, s);
-        })();
-    </script>
-
-</head>
-<body class="step1">
-
 <div class="container">
     <div class="alert alert-danger"><strong>Oops!</strong> <span></span></div>
     <div class="jumbotron">
@@ -166,6 +133,3 @@
 <script src="js/vendor/glfx.min.js"></script>
 
 <script src="js/main.js"></script>
-    
-</body>
-</html>
